Only try to cleanup our temp directory on Python 2, since Python 3 does it in some cases, and crashes if it's already clean.  Fixes #154

+++ green/process.py
@@ -254,7 +254,8 @@ def poolRunner(target, queue, coverage_number=None, omit_patterns=[]):  # pragma
 
     def cleanup():
         # Restore the state of the temp directory
-        shutil.rmtree(tempfile.tempdir, ignore_errors=True)
+        if sys.version_info[0] == 2: # pragma: no cover
+            shutil.rmtree(tempfile.tempdir, ignore_errors=True)
         tempfile.tempdir = saved_tempdir
         queue.put(None)
         # Finish coverage
